Refactor: 커스텀 hooks
